api/src/main/kotlin/org/burbridge/sandbox/api/repository/core/UserRepository.kt
package org.burbridge.sandbox.api.repository.core

import org.burbridge.sandbox.api.domain.core.User
import org.springframework.data.jpa.repository.JpaRepository
import org.springframework.data.jpa.repository.Modifying
import org.springframework.data.jpa.repository.Query
import org.springframework.data.repository.query.Param
import org.springframework.stereotype.Repository
import org.springframework.transaction.annotation.Transactional
import java.util.*

@Repository
interface UserRepository : JpaRepository<User, Long> {
    fun findByEmail(email: String): User?

    @Query("UPDATE User u SET u.lastLogin=:lastLogin WHERE u.email = ?#{ principal?.username }")
    @Modifying
    @Transactional
    fun updateLastLogin(@Param("lastLogin") lastLogin: Date)
}